🚨 Extract `entities` from `tree_db.root_paths`
TG-2031
