+++ src/youwol/backends/tree_db/routers/__init__.py
@@ -1,0 +1,2 @@
+# relative
+from .entities import router as router_entities
